# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Business Automation SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentRequest"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{departmentId}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get department
      security:
        - bearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    put:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      security:
        - bearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartmentRequest"
      responses:
        "200":
          description: Department updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete department
      security:
        - bearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /billing/subscriptions:
    get:
      operationId: listSubscriptions
      tags: [billing]
      summary: List subscriptions (Super Admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [billing]
      summary: Create / update subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /billing/invoices:
    get:
      operationId: listInvoices
      tags: [billing]
      summary: List invoices
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
  /marketing/campaigns:
    get:
      operationId: listCampaigns
      tags: [marketing]
      summary: List marketing campaigns
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [marketing]
      summary: Create campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /marketing/campaigns/{campaignId}:
    get:
      operationId: getCampaign
      tags: [marketing]
      summary: Get campaign
      security:
        - bearerAuth: []
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    put:
      operationId: updateCampaign
      tags: [marketing]
      summary: Update campaign
      security:
        - bearerAuth: []
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignRequest"
      responses:
        "200":
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [marketing]
      summary: Delete campaign
      security:
        - bearerAuth: []
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Department:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        headUserId:
          type: integer
          nullable: true
        headName:
          type: string
          nullable: true
        employeeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - employeeCount
        - createdAt
    CreateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        headUserId:
          type: integer
          nullable: true
      required:
        - name
    UpdateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        headUserId:
          type: integer
          nullable: true
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        tenantName:
          type: string
        planId:
          type: integer
        planName:
          type: string
        status:
          type: string
          enum: [active, past_due, cancelled, trialing]
        currentPeriodStart:
          type: string
          format: date-time
        currentPeriodEnd:
          type: string
          format: date-time
        amount:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - tenantName
        - planId
        - planName
        - status
        - currentPeriodStart
        - currentPeriodEnd
        - amount
        - createdAt
    CreateSubscriptionRequest:
      type: object
      properties:
        tenantId:
          type: integer
        planId:
          type: integer
      required:
        - tenantId
        - planId
    Invoice:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        tenantName:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [draft, open, paid, void, uncollectible]
        dueDate:
          type: string
          format: date
        paidAt:
          type: string
          format: date-time
          nullable: true
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - tenantName
        - amount
        - status
        - dueDate
        - description
        - createdAt
    Campaign:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [email, sms, social, push]
        status:
          type: string
          enum: [draft, scheduled, running, paused, completed]
        subject:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        targetAudience:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        sentCount:
          type: integer
        openCount:
          type: integer
        clickCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - type
        - status
        - sentCount
        - openCount
        - clickCount
        - createdAt
    CreateCampaignRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [email, sms, social, push]
        subject:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        targetAudience:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
      required:
        - name
        - type
    UpdateCampaignRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [draft, scheduled, running, paused, completed]
        subject:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        targetAudience:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
...